Backend/requirements.txt
python-dotenv
pymongo
beautifulsoup4
requests
rake-nltk